Big Changes read description
Models now when load, just get added to an array

At render time, we extract and apply the model info, instead of doing a global array.

So much easier to understand and makes more sense when I think of it now lol.

Basicall instead of drawing all the models based on vertices, we draw them per model